re-designed the codebase in cleaner way and updated the readme.md
