+++ url_shortner_app/src/main.rs
@@ -1,28 +1,10 @@
+mod server_service;
+use server_service::shortner::url_shortner_service_server::UrlShortnerServiceServer;
+use server_service::UrlShortnerServer;
+
 
 use tonic::{Request, Response, Status};
 use tonic::transport::Server;
-use shortner::url_shortner_service_server::{UrlShortnerService, UrlShortnerServiceServer};
-use shortner::{CreateShortenUrlPayload, Shorten}; // these messages acts like structs
-
-pub mod shortner {
-    tonic::include_proto!("url_shortner");
-}
-
-#[derive(Debug, Default)]
-pub struct UrlShortnerServer {}
-
-#[tonic::async_trait]
-impl UrlShortnerService for UrlShortnerServer{
-    async fn create_shorten_url(&self, request: Request<CreateShortenUrlPayload>) -> Result<Response<Shorten>, Status> {
-        let payload = request.into_inner();
-
-        let reply = Shorten {
-            shorten_url: payload.custom_url,
-        };
-
-        Ok(Response::new(reply))
-    }
-}
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
